Group the Nemotron-H patches into an architectures package (#651)
* Group the Nemotron-H patches into an architectures package

The Liger kernels and the Mamba2 mixer workarounds were split across
llm_ops and third_party_patches with no shared structure. They now sit under
agilerl/architectures/nemotron_h, behind a family registry that
install_zero3_third_party_hooks dispatches through, so a second architecture
is a new subpackage and a table entry.

third_party_patches keeps the generic ZeRO-3 patches. The resolution and
idempotence primitives move down to agilerl/utils/patching so the generic and
per-family modules can share them without importing each other.

No behaviour change: the same patches install at the same points for the same
checkpoints.

* Rename third_party_patches to zero3_patches

Everything left in the module is ZeRO-3: all three public symbols, both
resolved deepspeed classes and every trace constant. The third-party framing
belonged to the file when it also held the transformers-side Nemotron
patches, and the general notion now lives in utils/patching.

install_zero3_third_party_hooks loses the same vestigial word.

* [pre-commit.ci] auto fixes from pre-commit hooks

---------
