+++ pyproject.toml
@@ -189,7 +189,7 @@ include = ["agilerl", "agilerl/arena"]
 include = [
     "agilerl/algorithms/core/base.py",
     "agilerl/algorithms/core/llm_ops/fused_loss.py",
-    "agilerl/algorithms/core/llm_ops/liger_nemotron_h.py",
+    "agilerl/architectures/nemotron_h/liger.py",
     "agilerl/algorithms/core/llm_ops/moe_lora.py",
     "agilerl/algorithms/core/llm_ops/vllm_colocate.py",
     "agilerl/algorithms/grpo.py",
